Refactor code for ND proofs.
